Konstruktor für Karte angelegt Karten als Instanzen angelegt in Klasse Karten
